Added Airbnb swift lint / format rules
Run the lint command with
```sh
./lint.sh
```

This will automatically download / install the swiftlint and format deps
to run the lint config. Changes will be automatically added with `git
add .`. A portion of the scripts also changes the xcode editor settings
via CLI commands to not conflict with the formatting rules (ex: changing
default indent level to 2 instead of 4 spaces). You must run this before committing things in the repo.